requirements.txt
fyers-apiv3>=3.1.0
yfinance>=0.2.40
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
pyotp>=2.9.0
python-dotenv>=1.0.0
pytz>=2024.1
